phish-inference/script_requirements.txt
pandas==2.0.1
nostril==0.1.1
Levenshtein==0.21.0
requests>=2.16.0
tldextract==3.4.1
#urllib
beautifulsoup4>=4.1
pyyaml>=6.0